Eliminated a number of warnings about incorrect printf() formatting directives.

+++ cel/trunk/plugins/behaviourlayer/xml/xmlscript.cpp
@@ -2480,9 +2480,9 @@ bool celXmlScriptEventHandler::Execute (iCelEntity* entity,
 	  	A2S (val)));
 	  if (varprop_trace)
 	  {
-	    printf (":%s/%04d: lvar varidx=%s value=%s\n",
+	    printf (":%s/%04lu: lvar varidx=%s value=%s\n",
 	    	cbl->call_stack.Top (),
-	    	i-1, A2S (op.arg), A2S (val));
+	        (unsigned long)(i-1), A2S (op.arg), A2S (val));
 	    fflush (stdout);
 	  }
 	  celXmlArg& lv = local_vars.GetExtend (op.arg.arg.ui);
@@ -2781,9 +2781,9 @@ bool celXmlScriptEventHandler::Execute (iCelEntity* entity,
           DUMP_EXEC ((":%04d: createpropclass %s\n", i-1, A2S (top)));
 	  if (varprop_trace)
 	  {
-	    printf (":%s/%04d: createpropclass %s\n",
+	    printf (":%s/%04lu: createpropclass %s\n",
 	    	cbl->call_stack.Top (),
-	    	i-1, A2S (top));
+		(unsigned long)i-1, A2S (top));
 	    fflush (stdout);
 	  }
 	  const char* s = ArgToString (top);
@@ -2803,9 +2803,9 @@ bool celXmlScriptEventHandler::Execute (iCelEntity* entity,
 	  	A2S (aent), A2S (abh)));
 	  if (varprop_trace)
 	  {
-	    printf (":%s/%04d: createentity %s behaviour=%s\n",
+	    printf (":%s/%04lu: createentity %s behaviour=%s\n",
 	    	cbl->call_stack.Top (),
-	    	i-1, A2S (aent), A2S (abh));
+		(unsigned long)i-1, A2S (aent), A2S (abh));
 	    fflush (stdout);
 	  }
 	  csRef<iCelEntity> ent = pl->CreateEntity ();
@@ -3029,9 +3029,9 @@ bool celXmlScriptEventHandler::Execute (iCelEntity* entity,
 	  DUMP_EXEC ((":%04d: destroyentity %s\n", i-1, A2S (aent)));
 	  if (varprop_trace)
 	  {
-	    printf (":%s/%04d: destroyentity %s\n",
+	    printf (":%s/%04lu: destroyentity %s\n",
 	    	cbl->call_stack.Top (),
-	    	i-1, A2S (aent));
+		(unsigned long)i-1, A2S (aent));
 	    fflush (stdout);
 	  }
 	  iCelEntity* ent = ArgToEntity (aent, pl);
@@ -3516,9 +3516,9 @@ bool celXmlScriptEventHandler::Execute (iCelEntity* entity,
 	  	A2S (a_ent), A2S (var), A2S (val)));
 	  if (varprop_trace)
 	  {
-	    printf (":%s/%04d: varent ent=%s var=%s value=%s\n",
+	    printf (":%s/%04lu: varent ent=%s var=%s value=%s\n",
 	    	cbl->call_stack.Top (),
-	    	i-1, A2S (a_ent), A2S (var), A2S (val));
+		(unsigned long)i-1, A2S (a_ent), A2S (var), A2S (val));
 	    fflush (stdout);
 	  }
 
@@ -3549,9 +3549,9 @@ bool celXmlScriptEventHandler::Execute (iCelEntity* entity,
 	  	A2S (val)));
 	  if (varprop_trace)
 	  {
-	    printf (":%s/%04d: var_str var=%s value=%s\n",
+	    printf (":%s/%04lu: var_str var=%s value=%s\n",
 	    	cbl->call_stack.Top (),
-	    	i-1, A2S (op.arg), A2S (val));
+		(unsigned long)i-1, A2S (op.arg), A2S (val));
 	    fflush (stdout);
 	  }
 
@@ -3573,9 +3573,9 @@ bool celXmlScriptEventHandler::Execute (iCelEntity* entity,
 	  	A2S (val)));
 	  if (varprop_trace)
 	  {
-	    printf (":%s/%04d: var var=%s value=%s\n",
+	    printf (":%s/%04lu: var var=%s value=%s\n",
 	    	cbl->call_stack.Top (),
-	    	i-1, A2S (var), A2S (val));
+		(unsigned long)i-1, A2S (var), A2S (val));
 	    fflush (stdout);
 	  }
 
@@ -3626,9 +3626,9 @@ bool celXmlScriptEventHandler::Execute (iCelEntity* entity,
 	  DUMP_EXEC ((":%04d: inventory_add ent=%s\n", i-1, A2S (a_ent)));
 	  if (varprop_trace)
 	  {
-	    printf (":%s/%04d: inventory_add ent=%s\n",
+	    printf (":%s/%04lu: inventory_add ent=%s\n",
 	    	cbl->call_stack.Top (),
-	    	i-1, A2S (a_ent));
+		(unsigned long)i-1, A2S (a_ent));
 	    fflush (stdout);
 	  }
 
@@ -3649,9 +3649,9 @@ bool celXmlScriptEventHandler::Execute (iCelEntity* entity,
 	  DUMP_EXEC ((":%04d: inventory_rem ent=%s\n", i-1, A2S (a_ent)));
 	  if (varprop_trace)
 	  {
-	    printf (":%s/%04d: inventory_rem ent=%s\n",
+	    printf (":%s/%04lu: inventory_rem ent=%s\n",
 	    	cbl->call_stack.Top (),
-	    	i-1, A2S (a_ent));
+		(unsigned long)i-1, A2S (a_ent));
 	    fflush (stdout);
 	  }
 
@@ -3719,9 +3719,9 @@ bool celXmlScriptEventHandler::Execute (iCelEntity* entity,
 	  	A2S (a_id), A2S (a_val)));
 	  if (varprop_trace)
 	  {
-	    printf (":%s/%04d: property pc=%s id=%s val=%s\n",
+	    printf (":%s/%04lu: property pc=%s id=%s val=%s\n",
 	    	cbl->call_stack.Top (),
-	    	i-1, A2S (a_pc), A2S (a_id), A2S (a_val));
+		(unsigned long)i-1, A2S (a_pc), A2S (a_id), A2S (a_val));
 	    fflush (stdout);
 	  }
 	  iCelPropertyClass* pc = ArgToPClass (a_pc);
